pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nomad-commerce-analytics"
version = "0.1.0"
description = "Analytics Engineering demo with real (Olist) + synthetic data: DuckDB + dbt + Streamlit."
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Your Name" }]
license = { text = "MIT" }
dependencies = [
  # Core data + DB
  "duckdb>=0.9.2",
  "pandas>=2.1",
  "pyarrow>=14",
  "sqlalchemy>=2.0",
  # dbt + DuckDB adapter
  "dbt-core~=1.7",
  "dbt-duckdb~=1.7",
  # App
  "streamlit>=1.25",
  "altair>=5.0",
  # Analytics / stats
  "numpy>=1.26",
  "scipy>=1.11",
  "statsmodels>=0.14",
  # Geospatial basics
  "haversine>=2.8.0",
  # CLI/UX
  "typer>=0.9",
  # Config
  "pyyaml>=6.0",
  # Caching
  "diskcache>=5.6.3"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "pytest-cov>=4.1",
  "hypothesis>=6.88",
  "black>=23.7",
  "ruff>=0.1.13",
  "pre-commit>=3.3",
  "sqlfluff>=2.3",
  "sqlfluff-templater-dbt>=2.3",
  "detect-secrets>=1.4",
  "ipykernel>=6.25"
]

[tool.black]
line-length = 100
target-version = ["py310"]
include = "\\.pyi?$"
exclude = """
/(
  \\.git
| \\.venv
| build
| dist
| site
)/
"""

[tool.ruff]
line-length = 100
target-version = "py310"
extend-select = ["B", "C4", "I", "N", "PIE", "PT", "RUF", "SIM"]
ignore = [
  "E501"  # handled by black's wrapping
]
exclude = ["site/"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-q --disable-warnings"
testpaths = ["tests/python", "tests/e2e"]
filterwarnings = [
  "ignore::UserWarning",
]

[tool.coverage.run]
branch = true
source = ["scripts", "app"]
omit = ["site/*", "tests/*"]

[tool.coverage.report]
fail_under = 80
skip_empty = true

[tool.setuptools.packages.find]
where = ["."]
exclude = ["site", "data", "warehouse", "artifacts", "exports"]


